déplace dans un fichier toutes les méthodes liées à la synchronisation des logs
