+++ pkg/stats/sql.go
@@ -3,11 +3,8 @@ package stats
 import (
 	"context"
 	_ "embed"
-	"time"
 
-	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/pkg/errors"
 
 	"datapi/pkg/core"
 )
@@ -37,39 +34,4 @@ func getLastAccessLog(ctx context.Context, db *pgxpool.Pool) (core.AccessLog, er
 	row := db.QueryRow(ctx, "SELECT path, method, body, token FROM logs ORDER BY date_add DESC LIMIT 1;")
 	err := row.Scan(&r.Path, &r.Method, &r.Body, &r.Token)
 	return r, err
-}
-
-func syncAccessLogs(ctx context.Context, source *pgxpool.Pool, target *pgxpool.Pool) (int, error) {
-	rows, err := source.Query(ctx, `select id, date_add, path, method, body, token from logs`)
-	if err != nil {
-		return -1, errors.Wrap(err, "erreur pendant la récupération des logs depuis la source")
-	}
-	var batch pgx.Batch
-	for rows.Next() {
-		var currentID int
-		var currentTime time.Time
-		var currentLog core.AccessLog
-		err := rows.Scan(&currentID, &currentTime, &currentLog.Path, &currentLog.Method, &currentLog.Body, &currentLog.Token)
-		if err != nil {
-			return -1, errors.Wrap(err, "erreur pendant la lecture d'une log")
-		}
-		batch.Queue(`insert into logs (id, date_add, path, method, body, token) values ($1, $2 ,$3, $4, $5, $6)`,
-			currentID, currentTime, currentLog.Path, currentLog.Method, string(currentLog.Body), currentLog.Token)
-	}
-	br := target.SendBatch(ctx, &batch)
-	_, err = br.Exec()
-	if err != nil {
-		return -1, errors.Wrap(err, "erreur pendant l'insertion logs vers la cible")
-	}
-	return batch.Len(), nil
-}
-
-func countAccessLogs(ctx context.Context, db *pgxpool.Pool) (int, error) {
-	var counter int
-	row := db.QueryRow(ctx, "SELECT COUNT(*) FROM logs")
-	err := row.Scan(&counter)
-	if err != nil {
-		return -1, errors.Wrap(err, "erreur pendant le comptage des access logs")
-	}
-	return counter, nil
 }
