+++ pkg/stats/sql/create_or_replace.sql
@@ -3,7 +3,7 @@
 --
 
 CREATE SEQUENCE IF NOT EXISTS logs_id
-  START WITH 999999
+  START WITH 1
   INCREMENT BY 1
   NO MINVALUE
   NO MAXVALUE
